requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
djangorestframework==3.15.1
idna==3.6
newsapi-python==0.2.7
pillow==10.3.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
